Refactor localization provider for ui_locales support
Removed the `QueryStringToCookieRequestCultureProvider` class and introduced `UiLocalesRequestCultureProvider` to handle localization based on the `ui_locales` query string parameter. The new implementation includes methods for culture determination, instance creation, and value parsing, while maintaining cookie management logic. Additional using directives were added to support the new functionality.
